move enums to library root. continuing WIP to troubleshoot deserialization of bitfields.
