+++ src/enet.rs
@@ -2,7 +2,7 @@
 //!
 //! Diagram at https://en.wikipedia.org/wiki/Ethernet_frame#Ethernet_II
 
-use crate::MacAddr;
+use crate::{MacAddr, EtherType};
 
 use byte_struct::*;
 
@@ -92,75 +92,6 @@ where
         let mut bytes = [0_u8; Self::BYTE_LEN];
         self.write_bytes(&mut bytes);
         bytes
-    }
-}
-
-/// EtherType tag values (incomplete list - there are many more not implemented here)
-///
-/// See https://en.wikipedia.org/wiki/EtherType
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
-#[repr(u16)]
-pub enum EtherType {
-    /// IPV4
-    IPV4 = 0x0800,
-    /// ARP
-    ARP = 0x0806,
-    /// VLAN - if this tag is encountered, then this is not the real ethertype field, and we're reading an 802.1Q VLAN tag instead
-    VLAN = 0x8100,
-    /// IPV6
-    IPV6 = 0x86DD,
-    /// EtherCat
-    EtherCat = 0x88A4,
-    /// Precision Time Protocol
-    PTP = 0x88A7,
-    /// Catch-all for uncommon types not handled here
-    Unimplemented,
-}
-
-impl From<u16> for EtherType {
-    fn from(value: u16) -> Self {
-        match value {
-            x if x == EtherType::ARP as u16 => EtherType::ARP,
-            x if x == EtherType::EtherCat as u16 => EtherType::EtherCat,
-            x if x == EtherType::IPV4 as u16 => EtherType::IPV4,
-            x if x == EtherType::IPV6 as u16 => EtherType::IPV6,
-            x if x == EtherType::PTP as u16 => EtherType::PTP,
-            x if x == EtherType::VLAN as u16 => EtherType::VLAN,
-            _ => EtherType::Unimplemented,
-        }
-    }
-}
-
-impl ByteStructLen for EtherType {
-    const BYTE_LEN: usize = 2;
-}
-
-impl ByteStruct for EtherType {
-    fn read_bytes(bytes: &[u8]) -> Self {
-        if bytes.len() < 2 {
-            return EtherType::Unimplemented;
-        } else {
-            let mut bytes_read = [0_u8; 2];
-            bytes_read.copy_from_slice(&bytes[0..=1]);
-            return EtherType::from(u16::from_be_bytes(bytes_read));
-        }
-    }
-
-    fn write_bytes(&self, bytes: &mut [u8]) {
-        if bytes.len() >= 2 {
-            let bytes_to_write = (*self as u16).to_be_bytes();
-            bytes[0] = bytes_to_write[0];
-            bytes[1] = bytes_to_write[1];
-        } else {
-            // Do nothing - no bytes to write
-        }
-    }
-}
-
-impl EtherType {
-    /// Pack into big-endian (network) byte array
-    pub fn to_be_bytes(&self) -> [u8; Self::BYTE_LEN] {
-        (*self as u16).to_be_bytes()
     }
 }
